last one may be

+++ src/Pages/Admin/VehiclesListPage/VehiclesListPage.jsx
@@ -22,8 +22,11 @@ const VehiclesListPage = () => {
 
 
 const deleteVehicle = async (id) => {
+    const res = confirm("Are you sure! you want to delete")
+
+    if(res){
         try {
-            await fetch(`http://localhost:5000/api/vehicle/delete/${id}`,{
+            await fetch(`${import.meta.env.VITE_URL}/api/vehicle/delete/${id}`,{
                 method: "DELETE"
             })
             toast.success('Vehicle Deleted Successfully!', {
@@ -49,7 +52,19 @@ const deleteVehicle = async (id) => {
                 });
         }
         
+    }else{
+        toast.error('Request Cancel!', {
+            position: "top-center",
+            autoClose: 5000,
+            hideProgressBar: false,
+            closeOnClick: false,
+            pauseOnHover: true,
+            draggable: true,
+            progress: undefined,
+            theme: "light",
+            });
     }
+}
 
     // const updateVehicleStatus = async (vehicle) => {
     //     const id = vehicle._id;
@@ -57,7 +72,7 @@ const deleteVehicle = async (id) => {
     //         alert("Vehicle is already Available!")
     //     }else{
     //     try {
-    //         await fetch(`http://localhost:5000/api/vehicle/updatestatus/${id}`,{
+    //         await fetch(`${import.meta.env.VITE_URL}/api/vehicle/updatestatus/${id}`,{
     //             method: "PUT"
     //         })
     //         toast.success('Vehicle Deleted Successfully!', {
@@ -86,6 +101,22 @@ const deleteVehicle = async (id) => {
         
     // }
 
+    const [searchQuery, setSearchQuery] = useState("");
+        
+            // Filtering function
+            const filteredVehicles = vehicledata && vehicledata.filter(vehicle => {
+                const query = searchQuery.toLowerCase();
+                return (
+                    vehicle.user_id.toLowerCase().includes(query) ||
+                    vehicle.user_name.toLowerCase().includes(query) ||
+                    vehicle.Vehicle_type.toLowerCase().includes(query) ||
+                    vehicle.Vehicle_name.toLowerCase().includes(query) ||
+                    vehicle.Vehicle_model.toLowerCase().includes(query) ||
+                    vehicle.Vehicle_rent.toLowerCase().includes(query) ||
+                    vehicle.Vehicle_average.toLowerCase().includes(query)
+                    // vehicle.isBooked.includes(query)
+                );
+            });
 
   return (
     <>
@@ -100,6 +131,10 @@ const deleteVehicle = async (id) => {
               setSidebarVisiblity={setSidebarVisiblity}
               sidebarVisiblity={sidebarVisiblity} />
               <div className={Styles.bookingtable}>
+              <div style={{width: "95%", margin: "10px auto"}}>
+                    <label htmlFor="">Search:</label> <br />
+                    <input onChange={(e)=> setSearchQuery(e.target.value)} placeholder='Search' type="text" name="" id="" />
+                </div>
               <table>
                     <tr>
                         <th>S.No</th>
@@ -113,7 +148,7 @@ const deleteVehicle = async (id) => {
                     </tr>
                     
                     {
-                        vehicledata && vehicledata.map((vehicle, i) => {
+                        filteredVehicles?.length ? filteredVehicles.map((vehicle, i) => {
                             return(
                                 <tr key={vehicle._id}>
                                 <td>{i+1}</td>
@@ -121,7 +156,7 @@ const deleteVehicle = async (id) => {
                                 <td title={vehicle._id}>{vehicle._id}</td>
                                 <td>{vehicle.user_name}</td>
                                 <td>
-                                    <img src={`http://localhost:5000${vehicle.Vehicle_image}`} alt="" />
+                                    <img src={`${import.meta.env.VITE_URL}${vehicle.Vehicle_image}`} alt="" />
                                     <p>{vehicle.Vehicle_name}</p>
                                     {/* <p>025414</p> */}
                                 </td>
@@ -133,6 +168,8 @@ const deleteVehicle = async (id) => {
                             </tr>
                             )
                         })
+                        :
+                        <h1 style={{margin: "20px auto", textAlign: 'center', textTransform: 'uppercase', color:'#0061ff'}}>No Vehicles Found!</h1>
                     }
